v0.12.5: - Switched CUDA C development to CLion - Used the __JETBRAINS_IDE__ macro to declare CUDA's language extensions - Header include paths are now relative to the given source file - For real-time compilation with nvrtc, all include directives in the source code are replaced with a sequence of directives that use paths relative to the CUDA resource base directory. - Header files are now inferred from the source code and do no longer have to be specified in kernel instructions. - Fixed comparisons in the binary testing kernel - Replaced double constants with floats - Removed the unused numberEntries parameter from the kernel that replaces NaNs - Removed unused parameter from functions used for backpropagation kernels of recurrent layers - Resolved a name conflict in the max-pooling kernel - Simplified the definition of the stack of convolutional layers in the embedding toy demo with two filter widths

+++ src/main/resources/cuda/optimization/adaptive/AdamKernel.cu
@@ -1,3 +1,5 @@
+#include "../../cuda.h"
+
 __global__ void adamKernel (
     int numberIterations,
     int* parameterIndices,
@@ -21,7 +23,7 @@ __global__ void adamKernel (
 
     if(parameterIndex != -1 && count > 0) {
 
-        float scalingFactor = 1.0 / (float)count;
+        float scalingFactor = 1.0f / (float)count;
 
         int startEntryIndex = (blockIdx.y * blockDim.x + threadIdx.x) * numberIterations;
 
@@ -40,11 +42,11 @@ __global__ void adamKernel (
 
             float updatedFirstMomentEstimate = firstMomentDecay * firstMomentEstimate[parameterEntryIndex] + oneMinusFirstMomentDecay * scaledDerivative;
             firstMomentEstimate[parameterEntryIndex] = updatedFirstMomentEstimate;
-            float correctedFirstMomentEstimate = updatedFirstMomentEstimate / (1.0 - powf(firstMomentDecay, step));
+            float correctedFirstMomentEstimate = updatedFirstMomentEstimate / (1.0f - powf(firstMomentDecay, step));
 
             float updatedSecondMomentEstimate = secondMomentDecay * secondMomentEstimate[parameterEntryIndex] + oneMinusSecondMomentDecay * scaledDerivative * scaledDerivative;
             secondMomentEstimate[parameterEntryIndex] = updatedSecondMomentEstimate;
-            float correctedSecondMomentEstimate = updatedSecondMomentEstimate / (1.0 - pow(secondMomentDecay, step));
+            float correctedSecondMomentEstimate = updatedSecondMomentEstimate / (1.0f - powf(secondMomentDecay, step));
 
             float adaptedLearningRate = learningRate / (sqrtf(correctedSecondMomentEstimate) + epsilon);
